+++ tests/functional/merge/cram/merge.t
@@ -79,6 +79,26 @@ Supports --metadata-id-columns.
   two    X2a  X2b  Y2c  Y2f  Y2e  Y2d                    1                    1
   three                 Y3f  Y3e  Y3d                    0                    1
 
+Supports table-specific --metadata-id-columns.
+
+  $ ${AUGUR} merge \
+  >   --metadata X=x-id-column.tsv Y=y.tsv \
+  >   --metadata-id-columns X=id \
+  >   --output-metadata - --quiet | csv2tsv --csv-delim $'\t' | tsv-pretty
+  id     a    b    c    f    e    d    __source_metadata_X  __source_metadata_Y
+  one    X1a  X1b  X1c                                   1                    0
+  two    X2a  X2b  Y2c  Y2f  Y2e  Y2d                    1                    1
+  three                 Y3f  Y3e  Y3d                    0                    1
+
+  $ ${AUGUR} merge \
+  >   --metadata X=x.tsv Y=y-name-column.tsv \
+  >   --metadata-id-columns strain Y=name \
+  >   --output-metadata - --quiet | csv2tsv --csv-delim $'\t' | tsv-pretty
+  strain  a    b    c    f    e    d    __source_metadata_X  __source_metadata_Y
+  one     X1a  X1b  X1c                                   1                    0
+  two     X2a  X2b  Y2c  Y2f  Y2e  Y2d                    1                    1
+  three                  Y3f  Y3e  Y3d                    0                    1
+
 Supports Augur's standard delimiter detection.
 
   $ sed 's/\t/,/g' < x.tsv > x.csv
@@ -96,6 +116,26 @@ Supports --metadata-delimiters.
   $ ${AUGUR} merge \
   >   --metadata X=x.txt Y=y.tsv \
   >   --metadata-delimiters '|' $'\t' \
+  >   --output-metadata - --quiet | csv2tsv --csv-delim $'\t' | tsv-pretty
+  strain  a    b    c    f    e    d    __source_metadata_X  __source_metadata_Y
+  one     X1a  X1b  X1c                                   1                    0
+  two     X2a  X2b  Y2c  Y2f  Y2e  Y2d                    1                    1
+  three                  Y3f  Y3e  Y3d                    0                    1
+
+Supports table-specific --metadata-delimiters.
+
+  $ ${AUGUR} merge \
+  >   --metadata X=x.txt Y=y.tsv \
+  >   --metadata-delimiters X='|' \
+  >   --output-metadata - --quiet | csv2tsv --csv-delim $'\t' | tsv-pretty
+  strain  a    b    c    f    e    d    __source_metadata_X  __source_metadata_Y
+  one     X1a  X1b  X1c                                   1                    0
+  two     X2a  X2b  Y2c  Y2f  Y2e  Y2d                    1                    1
+  three                  Y3f  Y3e  Y3d                    0                    1
+
+  $ ${AUGUR} merge \
+  >   --metadata X=x.txt Y=y.tsv \
+  >   --metadata-delimiters $'\t' X='|' \
   >   --output-metadata - --quiet | csv2tsv --csv-delim $'\t' | tsv-pretty
   strain  a    b    c    f    e    d    __source_metadata_X  __source_metadata_Y
   one     X1a  X1b  X1c                                   1                    0
@@ -218,6 +258,60 @@ Duplicates.
   Error: stepping, UNIQUE constraint failed: metadata_dups.strain (19)
   WARNING: Skipped deletion of */augur-merge-*.sqlite due to error, but you may want to clean it up yourself (e.g. if it's large). (glob)
   ERROR: sqlite3 invocation failed
+  [2]
+
+Unknown metadata names in --metadata-delimiters.
+
+  $ ${AUGUR} merge \
+  >   --metadata X=x.tsv Y=y.tsv \
+  >   --metadata-delimiters $'\t' whatsit=';' \
+  >   --output-metadata -
+  ERROR: Unknown metadata table name in --metadata-delimiters:
+  
+    'whatsit'
+  
+  This name does not appear in the NAME=FILE pairs given to --metadata.
+  
+  [2]
+
+  $ ${AUGUR} merge \
+  >   --metadata X=x.tsv Y=y.tsv \
+  >   --metadata-delimiters whatsit=';' whosit=, X=$'\t' Y=$'\t' \
+  >   --output-metadata -
+  ERROR: Unknown metadata table names in --metadata-delimiters:
+  
+    'whatsit'
+    'whosit'
+  
+  These names do not appear in the NAME=FILE pairs given to --metadata.
+  
+  [2]
+
+Unknown metadata names in --metadata-id-columns.
+
+  $ ${AUGUR} merge \
+  >   --metadata X=x.tsv Y=y.tsv \
+  >   --metadata-id-columns strain whatsit=id \
+  >   --output-metadata -
+  ERROR: Unknown metadata table name in --metadata-id-columns:
+  
+    'whatsit'
+  
+  This name does not appear in the NAME=FILE pairs given to --metadata.
+  
+  [2]
+
+  $ ${AUGUR} merge \
+  >   --metadata X=x.tsv Y=y.tsv \
+  >   --metadata-id-columns whatsit=id whosit=accession X=strain Y=strain \
+  >   --output-metadata -
+  ERROR: Unknown metadata table names in --metadata-id-columns:
+  
+    'whatsit'
+    'whosit'
+  
+  These names do not appear in the NAME=FILE pairs given to --metadata.
+  
   [2]
 
 No id column found.